ibc_clients::tendermint::client_state
Function update_on_recovery

pub fn update_on_recovery<E>(
    subject_client_state: ClientState,
    ctx: &mut E,
    subject_client_id: &ClientId,
    substitute_client_state: Any,
    substitute_consensus_state: Any,
) -> Result<(), ClientError>
where
    E: ExtClientExecutionContext,
    <E as ClientValidationContext>::ClientStateRef: From<ClientState>,
    ConsensusState: Convertible<<E as ClientValidationContext>::ConsensusStateRef>,

Update the client_state’s ID, trusting period, latest height, processed height, and processed time metadata values to those values provided by a verified substitute client state in response to a successful client recovery.

Note that unlike the update_on_upgrade function, update_on_recovery assumes that the client being updated has already been re-initialised such that its original client and consensus states have been overwritten to their new states.

This function is typically implemented as part of the ClientStateExecution trait, but has been made standalone in order to enable greater flexibility of the ClientState APIs.