ibc_clients::tendermint::client_state
Function update_on_upgrade

pub fn update_on_upgrade<E>(
    client_state: &ClientState,
    ctx: &mut E,
    client_id: &ClientId,
    upgraded_client_state: Any,
    upgraded_consensus_state: Any,
) -> Result<Height, ClientError>
where
    E: ExtClientExecutionContext,
    <E as ClientValidationContext>::ClientStateRef: From<ClientState>,
    ConsensusState: Convertible<<E as ClientValidationContext>::ConsensusStateRef>,

Commit the new client state and consensus state to the store upon a successful client upgrade.

Note that this function is typically implemented as part of the ClientStateExecution trait, but has been made a standalone function in order to make the ClientState APIs more flexible.