ibc_clients::tendermint::client_state
Function verify_upgrade_client

pub fn verify_upgrade_client<H>(
    client_state: &ClientState,
    upgraded_client_state: Any,
    upgraded_consensus_state: Any,
    proof_upgrade_client: CommitmentProofBytes,
    proof_upgrade_consensus_state: CommitmentProofBytes,
    upgrade_path_prefix: CommitmentPrefix,
    upgrade_client_path_bytes: PathBytes,
    upgrade_consensus_path_bytes: PathBytes,
    root: &CommitmentRoot,
) -> Result<(), ClientError>
where
    H: HostFunctionsProvider,

Perform client-specific verifications and check all data in the new client state to be the same across all valid Tendermint clients for the new chain.

You can learn more about how to upgrade IBC-connected SDK chains in this guide.

Note that this function is typically implemented as part of the ClientStateCommon trait, but has been made a standalone function in order to make the ClientState APIs more flexible.